+++ src/wallet/receive.h
@@ -1,0 +1,20 @@
+// Copyright (c) 2021 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_WALLET_RECEIVE_H
+#define BITCOIN_WALLET_RECEIVE_H
+
+#include <amount.h>
+#include <wallet/ismine.h>
+#include <wallet/transaction.h>
+#include <wallet/wallet.h>
+
+struct COutputEntry
+{
+    CTxDestination destination;
+    CAmount amount;
+    int vout;
+};
+
+#endif // BITCOIN_WALLET_RECEIVE_H
